Add #hashByDivision and #hashByMultiplication methods to HashMap
These private methods will be used by the hash(key) method to take
into consideration the size of the buckets array.

They take the pure hash code computed converting the key to a number
and modify it in order to get a valid index in the buckets array,
returning an updated hash code.

The simpler #hashByDivision should use a prime buckets size, while
the size can be any number using #hashByMultiplication.
see https://www.geeksforgeeks.org/what-are-hash-functions-and-how-to-choose-a-good-hash-function/

Since the size of the buckets array is set as a multiple of 2, as
it is usually done, the #hashByMultiplication is actually used.

Also, to avoid numerical errors while iteratevly converting the
string to a number (the number might indeed quickly become very
large), we call #hashByMultiplication at each step of the process.