Setup the LinkedList class and use it for the buckets in HashMap
Each bucket in the #buckets array in the HashMap class is now a
LinkedList of Node objects, instead of a simple Node (holding a
key and a value).

Then, when setting some key and value, if there is a conflict (a
Node holding the same key is already present in the LinkedList in
the hash(key) bucket), just the value is updated (instead of
overwriting the whole Node).

If there is a collision (there is no Node holding the same key in
the LinkedList in the hash(key) bucket), a new Node is created and
appended.